pass only the property needed by the service

+++ src/main/kotlin/com/example/Properties.kt
@@ -4,5 +4,5 @@ import java.lang.System.getenv
 
 data class Properties(
     val serverPort: Int = 8080,
-    val exampleInitial: Int = (getenv("INITIAL") ?: "0").toInt()
+    val exampleInitial: Long = (getenv("INITIAL") ?: "0").toLong()
 )
